Give four feature modules the direct tests the mutation gate needs
The gate refuses to start on a source that has no test at its mirrored
path, so any branch touching these four was blocked however small the
change. Each now has one.

Scores after the pass: attendees-list 100%, attendee-page 98.1% (one
recorded equivalent), store-refund 96.6%, listing-page-data 77.8%. The
last two are not finished, and TODO.md records exactly what each
remaining survivor needs.

Along the way the filler-attendee seeder moved into test-utils, so the
new suite and the integration suite share one copy.
