+++ test/integration/server/attendees-list.test.ts
@@ -12,6 +12,7 @@ import { describeWithEnv } from "#test-utils/db.ts";
 import {
   createMultiBookingAttendee,
   createTestAttendeeDirect,
+  seedFillerAttendees,
 } from "#test-utils/db-helpers/attendees.ts";
 import {
   createTestListing,
@@ -134,36 +135,6 @@ describeWithEnv("server (admin attendees list)", { db: true }, () => {
      *  higher AUTOINCREMENT ids than every earlier row, so the newest-first
      *  (id-ordered) pagination under test sees them exactly like real
      *  bookings made after the "oldest" fixture attendee. */
-    const seedFillerAttendees = async (
-      listingId: number,
-      count: number,
-    ): Promise<void> => {
-      const { getDb } = await import("#shared/db/client.ts");
-      const { attendee } = await createTestAttendeeDirect(
-        listingId,
-        "Filler",
-        "filler@example.com",
-      );
-      const cloneStatements = [];
-      for (let i = 1; i < count; i++) {
-        cloneStatements.push(
-          {
-            args: [`filler-token-${i}`, attendee.id],
-            sql: `INSERT INTO attendees (created, kind, checked_in, ticket_token_index, pii_blob, status_id)
-                  SELECT created, kind, checked_in, ?, pii_blob, status_id
-                  FROM attendees WHERE id = ?`,
-          },
-          {
-            args: [attendee.id],
-            sql: `INSERT INTO listing_attendees (listing_id, attendee_id, start_at, end_at, quantity, checked_in)
-                  SELECT listing_id, last_insert_rowid(), start_at, end_at, quantity, checked_in
-                  FROM listing_attendees WHERE attendee_id = ?`,
-          },
-        );
-      }
-      await getDb().batch(cloneStatements, "write");
-    };
-
     test("shows the whole page with no paging links when the attendees exactly fill it", async () => {
       const listing = await makeListing("Full Page", ATTENDEES_PAGE_SIZE * 2);
       // Created first = oldest = the row a broken hasNext would trim off.
